Summary:
Resolves #4067. Performance-wise, this incurs some extra administrative
overhead on all invocations. In exchange, we can see order-of-magnitude
improvements in cases where there are expensive “fallback iteratees”
that are rarely hit (e.g., a `JSON.stringify`).

Running the `sortBy` benchmarks included in `perf/perf.js` gives the
following results on my laptop (Intel i5-4300U with Chrome 67 on Linux):

  - “`_.sortBy` with `callback`”:
      - before this patch: 59,249 ops/sec ±2.65% (56 runs sampled)
      - after this patch: 54,504 ops/sec ±2.91% (56 runs sampled)
      - (new / old) ratio: 0.92 ± 0.05 (0.87–0.97)

  - “`_.sortBy` with `property` name”:
      - before this patch: 98,956 ops/sec ±2.37% (58 runs sampled)
      - after this patch: 93,793 ops/sec ±1.46% (58 runs sampled)
      - (new / old) ratio: 0.94 ± 0.04 (0.91–0.98)

…and these on a workstation (Intel W-2135 with Chrome 71 on Linux):

  - “`_.sortBy` with `callback`”:
      - before this patch: 66,071 ops/sec ±2.39% (61 runs sampled)
      - after this patch: 58,707 ops/sec ±2.82% (58 runs sampled)
      - (new / old) ratio: 0.89 ± 0.05 (0.84–0.94)

  - “`_.sortBy` with `property` name”:
      - before this patch: 132,945 ops/sec ±2.71% (56 runs sampled)
      - after this patch: 129,162 ops/sec ±3.30% (61 runs sampled)
      - (new / old) ratio: 0.97 ± 0.06 (0.92–1.03)

(Best-of-3 timing used in each case.)

We can see that there is usually a statistically significant _decrease_
in the overall performance in these existing benchmarks.

This commit does not add a new benchmark to observe the best-case
behavior because the Underscore.js API does not support multiple
iteratees in `sortBy`, which would be needed for a fair comparison.

Test Plan:
The newly added test cases fail before this commit, but pass after it.
All existing tests pass (`npm run validate` passes).

wchargin-branch: sortBy-lazy