@wchargin 2019-01-07
@jdalton: Performance details in the description. There’s a reasonable
argument for not merging this or for making the behavior optional: it’s
up to you.