@jdalton 2019-07-09
Not sure. It wasn't an explicit close. Might be related to some branch activity going on. Fee free to resubmit the PR. The GitHub UI is not giving me the ability to Reopen it.